requirements.txt
pandas
kafka-python
sqlalchemy
psycopg2-binary
pyspark